[Refactor/#294] Merge CommentRequest and ReplyRequest into ConversationRequest
