+++ src/main/java/com/example/waggle/domain/conversation/persistence/dao/comment/querydsl/CommentQueryRepository.java
@@ -1,7 +1,7 @@
 package com.example.waggle.domain.conversation.persistence.dao.comment.querydsl;
 
-import com.example.waggle.domain.conversation.presentation.dto.comment.CommentResponse.QuestionCommentViewDto;
-import com.example.waggle.domain.conversation.presentation.dto.comment.CommentResponse.SirenCommentViewDto;
+import com.example.waggle.domain.conversation.presentation.dto.CommentResponse.QuestionCommentViewDto;
+import com.example.waggle.domain.conversation.presentation.dto.CommentResponse.SirenCommentViewDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
