--- src/main/java/com/example/waggle/domain/conversation/presentation/dto/reply/ReplyRequest.java
@@ -1,16 +1,0 @@
-package com.example.waggle.domain.conversation.presentation.dto.reply;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@Schema
-public class ReplyRequest {
-    private String content;
-}
